[Chore] 이미지 콜라쥬, 리사이징 extension 파일 분리
